requirements.txt
streamlit
langchain
langchain-openai
langchain-groq
mysql-connector-python 
python-dotenv